ModalManager – code that used to dismiss modals in Layout
