+++ app/src/marketplace/components/ErrorPageView/index.jsx
@@ -7,7 +7,7 @@ import { Translate, I18n } from 'react-redux-i18n'
 
 import BodyClass, { PAGE_SECONDARY } from '../BodyClass'
 import EmptyState from '$shared/components/EmptyState'
-import Layout from '../../containers/Layout'
+import Layout from '../Layout'
 import links from '../../../links'
 import appCrashedImage from '$shared/assets/images/app_crashed.png'
 import appCrashedImage2x from '$shared/assets/images/app_crashed@2x.png'
